Add script to train evidence networks  (#4)
* Add a configuration script

* Add evidence network training code
* Enable blind coverage test on trained networks 

* Add a models directory to store networks in
* Ignore en models in models directory for now

* Add figures directory to store figure in
* Ignore figures contents for now
* Add style file for matplotlib figures

* Update main .gitignore to include globalemu folders

* Bump version

+++ simulators/noise.py
@@ -35,7 +35,7 @@ def white_noise(shape: np.ndarray | int, sigma_noise: float) -> np.ndarray:
 
 def generate_white_noise_simulator(
         noise_size: int,
-        sigma_noise: int
+        sigma_noise: float
 ) -> Simulator:
     """Return a simulator function for white noise.
